requirements.txt
numpy>=1.16.0
Pillow>=9.0.0
matplotlib>=3.0.0
opencv-python>=4.5.0
tensorflow>=2.0.0
tqdm>=4.60.0
labelme>=4.5.0
pycocotools>=2.0.0
protobuf>=3.17.0
argparse>=1.4.0
scikit-image>=0.17.0
